move to static function
